### 2024-04-17 #6639 @editorialbot
[PRE REVIEW]: PyroRL: A Reinforcement Learning Environment for Wildfire Evacuation
**Submitting author:** @cpondoc (<a href="http://orcid.org/0009-0004-0317-5786">Christopher Pondoc</a>)
**Repository:** https://github.com/sisl/PyroRL
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.1
