+++ paper/paper.bib
@@ -45,11 +45,13 @@
     year = {2019}
 }
 
-@article{diao2020uncertainty,
-  title={Uncertainty aware wildfire management},
-  author={Diao, Tina and Singla, Samriddhi and Mukhopadhyay, Ayan and Eldawy, Ahmed and Shachter, Ross and Kochenderfer, Mykel},
-  journal={arXiv preprint arXiv:2010.07915},
-  year={2020}
+@InProceedings{Diao2020,
+author = {Tina Diao and Samriddhi Singla and Ayan Mukhopadhyay and Ahmed Eldawy and Ross Shachter and Mykel J. Kochenderfer},
+booktitle = {AIAA Fall Symposium},
+title = {Uncertainty aware wildfire management},
+year = {2020},
+url = {https://arxiv.org/pdf/2010.07915.pdf},
+doi = {10.48550/arXiv.2010.07915}
 }
 
 @article{ganapathi2018using,
@@ -59,18 +61,20 @@
   volume={5},
   pages={6},
   year={2018},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA}, 
+  doi = {10.3389/FICT.2018.00006}
 }
 
-@article{julian2019distributed,
-  title={{Distributed wildfire surveillance with autonomous aircraft using deep reinforcement learning}},
-  author={Julian, Kyle D and Kochenderfer, Mykel J},
-  journal={Journal of Guidance, Control, and Dynamics},
-  volume={42},
-  number={8},
-  pages={1768--1778},
-  year={2019},
-  publisher={American Institute of Aeronautics and Astronautics}
+@Article{Julian2019jgcdfire,
+author = {Kyle D. Julian and Mykel J. Kochenderfer},
+journal = jgcd,
+title = {Distributed wildfire surveillance with autonomous aircraft using deep reinforcement learning},
+year = {2019},
+number = {8},
+pages = {1768--1778},
+volume = {42},
+doi = {10.2514/1.G004106},
+url = {https://arxiv.org/abs/1810.04244},
 }
 
 @article{altamimi2022large,
@@ -80,7 +84,8 @@
   volume={5},
   pages={734330},
   year={2022},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi = {10.3389/ffgc.2022.734330}
 }
 
 @ARTICLE{9340340,
@@ -143,13 +148,19 @@
   publisher={MIT press}
 }
 
-@article{JMLR:v22:20-1364,
-  author  = {Antonin Raffin and Ashley Hill and Adam Gleave and Anssi Kanervisto and Maximilian Ernestus and Noah Dormann},
-  title   = {Stable-Baselines3: Reliable Reinforcement Learning Implementations},
-  journal = {Journal of Machine Learning Research},
-  year    = {2021},
-  volume  = {22},
-  number  = {268},
-  pages   = {1--8},
-  url     = {http://jmlr.org/papers/v22/20-1364.html}
+@article{10.5555/3546258.3546526,
+author = {Raffin, Antonin and Hill, Ashley and Gleave, Adam and Kanervisto, Anssi and Ernestus, Maximilian and Dormann, Noah},
+title = {Stable-baselines3: reliable reinforcement learning implementations},
+year = {2021},
+issue_date = {January 2021},
+publisher = {JMLR.org},
+volume = {22},
+number = {1},
+issn = {1532-4435},
[…]
