### 2024-04-24 #6639 @mikemahoney218
Happy to take this one on!

@cpondoc , would you be able to suggest 5 reviewers for this package? Reviewers can come from either [our reviewer database](https://reviewers.joss.theoj.org/) or your own network, so long as they don't have a [conflict of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy). When you suggest them, please _do not_ ping them using `@` -- we'll only do that when we're actually inviting them :smile:

After that, I should be able to start pinging people for reviews!
