+++ paper/paper.bib
@@ -148,19 +148,14 @@ url = {https://arxiv.org/abs/1810.04244},
   publisher={MIT press}
 }
 
-@article{10.5555/3546258.3546526,
-author = {Raffin, Antonin and Hill, Ashley and Gleave, Adam and Kanervisto, Anssi and Ernestus, Maximilian and Dormann, Noah},
-title = {Stable-baselines3: reliable reinforcement learning implementations},
-year = {2021},
-issue_date = {January 2021},
-publisher = {JMLR.org},
-volume = {22},
-number = {1},
-issn = {1532-4435},
-abstract = {STABLE-BASELINES3 provides open-source implementations of deep reinforcement learning (RL) algorithms in Python. The implementations have been benchmarked against reference codebases, and automated unit tests cover 95\% of the code. The algorithms follow a consistent interface and are accompanied by extensive documentation, making it simple to train and compare different RL algorithms.},
-journal = {J. Mach. Learn. Res.},
-month = {jan},
-articleno = {268},
-numpages = {8},
-keywords = {PyTorch, Python, open-source, software, baselines, reinforcement learning}
+@article{stableBaselines,
+  author  = {Antonin Raffin and Ashley Hill and Adam Gleave and Anssi Kanervisto and Maximilian Ernestus and Noah Dormann},
+  title   = {Stable-Baselines3: Reliable Reinforcement Learning Implementations},
+  journal = {Journal of Machine Learning Research},
+  year    = {2021},
+  volume  = {22},
+  number  = {268},
+  pages   = {1--8},
+  url     = {http://jmlr.org/papers/v22/20-1364.html},
+  doi = {10.5555/3546258.3546526}
 }
