### 2024-04-25 #6639 @mikemahoney218
Just in case it's helpful, people often want `@editorialbot generate pdf` rather than `generate preprint` -- `pdf` generates the formatted PDF file that you'd get after acceptance (and is persistent over time), while `preprint` generates TeX.
