### 2024-04-25 #6639 @cpondoc
Hey @mikemahoney218 — thanks again for the top about `generate pdf`! We’re still looking for reviewers, but I also wanted to update you on some of the DOI logistics. In particular:

- `Gym Cellular Automata` and `Gym Forest Fire` are just open-source packages/GitHub repositories with no literature associated with them
- `Algorithms for Decision Making` is a textbook with no associated DOI
- `A mathematical model…` is a paper from the USDA Forest Service from the 1970’s, so we also don’t believe there’s a DOI there
- The invalid DOI is for Stable-Baselines 3. Looking at the official link on the [ACM website](https://dl.acm.org/doi/abs/10.5555/3546258.3546526), the DOI we gave is what is in the URL as well as what is used by other papers that cite the authors. However, it seems to be marked as invalid (and it also seems to be undiscoverable from the POV of doi.org, as well)

Thanks again!
