### 2024-04-25 #6639 @mikemahoney218
Thanks! The Stable-baselines3 case is weird; JMLR doesn't assign DOIs (which is a weird decision in the first place) and it seems like ACM generates a "fake" DOI in its url, though if you click the "citation" button on that page or the [JMLR original](https://jmlr.org/papers/v22/20-1364.html), there's no DOI field in the generated bibtex. I think the best option would be to remove the DOI from that entry, as it's not a valid DOI.

Entries with no DOI are fine -- we want to make sure we're using DOIs wherever applicable, but we understand some resources don't. I haven't checked, but if you can make sure you've got the ISBN included for `Algorithms for Decision Making`, that'll be all set.
