### 2024-04-25 c7c315b
removed stable baselines doi and added text isbn
