+++ paper/paper.bib
@@ -145,7 +145,8 @@ url = {https://arxiv.org/abs/1810.04244},
   title={Algorithms for decision making},
   author={Kochenderfer, Mykel J and Wheeler, Tim A and Wray, Kyle H},
   year={2022},
-  publisher={MIT press}
+  publisher={MIT press},
+  isbn={0262047012}
 }
 
 @article{stableBaselines,
@@ -156,6 +157,5 @@ url = {https://arxiv.org/abs/1810.04244},
   volume  = {22},
   number  = {268},
   pages   = {1--8},
-  url     = {http://jmlr.org/papers/v22/20-1364.html},
-  doi = {10.5555/3546258.3546526}
+  url     = {http://jmlr.org/papers/v22/20-1364.html}
 }
