### 2024-04-26 #6639 @cpondoc
Sounds great! We just took out the DOI for Stable-baselines3 and also added the ISBN for `Algorithms for Decision Making` -- that should be everything on the references end!
