### 2024-05-22 daa49f7
Add logo to docs
