+++ docs/docs/index.md
@@ -1,4 +1,11 @@
-# PyroRL
+<pre>
+██████╗ ██╗   ██╗██████╗  ██████╗ ██████╗ ██╗     
+██╔══██╗╚██╗ ██╔╝██╔══██╗██╔═══██╗██╔══██╗██║     
+██████╔╝ ╚████╔╝ ██████╔╝██║   ██║██████╔╝██║     
+██╔═══╝   ╚██╔╝  ██╔══██╗██║   ██║██╔══██╗██║     
+██║        ██║   ██║  ██║╚██████╔╝██║  ██║███████╗
+╚═╝        ╚═╝   ╚═╝  ╚═╝ ╚═════╝ ╚═╝  ╚═╝╚══════╝
+</pre>
 
 PyroRL is a new reinforcement learning environment built for the simulation of wildfire evacuation.
 
