### 2024-05-23 sisl/PyroRL#12 @SamTov
Buggy pygame window on Mac Intel
I ran the `map_gen.py` example on an intel Mac with Python 3.10. The Pygame window is very glitchy, it seems to be reloading the entire window at every update.
