### 2024-05-23 sisl/PyroRL#13 @SamTov
Related work for paper review
The one thing missing for me is the related work. I did a very short search for RL and wildfire evacuation strategies and found a few papers:

* https://link.springer.com/article/10.1007/s10694-023-01363-1
* https://www.climatechange.ai/papers/neurips2021/22
* https://www.sciencedirect.com/science/article/abs/pii/S0378437121001175
* https://pubs.aip.org/aip/acp/article/3078/1/040005/3284720/A-reinforcement-learning-algorithm-for-the-optimal

and software

* https://github.com/elbecerrasoto/gym-forest-fire

I understand that these are  a bit different to what you have done but I guess it should be mentioned somewhere in the paper as it is a specific requirement in the JOSS guidelines.
