### 2024-05-26 sisl/PyroRL#12 @cpondoc
Hey @SamTov -- thanks for bringing this up! We'll investigate this on our end, but I will say that in terms of reloading the entire window at every update, that should be expected behavior. Specifically: calling `render` on our environment should save a snapshot at a singular point in time, so it would make sense that it reloads with each timestamp. Nonetheless, we'll report back!
