### 2024-05-27 #6739 @cpondoc
Hey @SamTov! Thanks for the comments -- we'll be sure to fix up the related work and also investigate the behavior on `map_gen_example.py` by the end of this week. Let us know if you need anything else from our end!
