### 2024-05-28 sisl/PyroRL#14 @shahchiragh
Nice to haves
Hi @cpondoc and team - Overall, the package looks great! I have a few nice-to-have suggestions for the package that can make any developer happy. I am listing them below. Please take a look and feel free to add them (optionally) as you find it suitable or convenient.

- Python version support. Adding a note what all versions the package works with would be a nice thing to have
- Requirements/ dependencies. When running  [stable_baselines.py](https://github.com/sisl/PyroRL/blob/master/examples/stable_baselines.py) directly, I encountered an error complaining about `No module named 'stable_baselines3'`. It would be helpful to have this requirement and any other requirements listed somewhere
- Example usage in readme. This is similar to a step-by-step walk-through on how to run `intro.py` from the examples module and pointing out the use of the `pyrorl` package. I understand this item is already shared in the demo video but it would be beneficial for other user audience who are more pivoted to reading the readme's
- [map_gen_example.py](https://github.com/sisl/PyroRL/blob/master/examples/map_gen_example.py) seems to be little quick on the reload making it visually hard to say whether its buggy or changing the images. May be slowing it down a bit might make it render compelling to the user
