### 2024-05-28 #6739 @mikemahoney218
Thank you @shahchiragh ! One quick question -- can you verify that you installed the software and got it to run, etc? I'm asking because it's a bit unusual to have reviews not require any issues being opened on the software repo. It happens occasionally, but just want to check in to make sure the process was clear to everyone given the circumstances. Thank you!
