### 2024-05-28 #6739 @mikemahoney218
Whoops -- I see https://github.com/sisl/PyroRL/issues/14 now; I didn't earlier. Apologies, and thank you so much for your review!
