### 2024-05-29 sisl/PyroRL#12 @SamTov
@cpondoc I should maybe be a little more specific as I am not sure if we are on the same page. When I say window, I mean the full GUI window, not just the frame of the simulation. From the examples I saw, this is not expected from your code. It looks like the gui is being shut down and re-loaded at each update, not just the frame.
