### 2024-08-08 5e37f95
Add a single line for quickstart in the README
