### 2024-08-08 68314a3
Specify python version across both docs and README
