+++ paper/paper.bib
@@ -9,6 +9,101 @@
     doi = {10.5281/zenodo.8127026},
 }
 
+@misc{tapley2023reinforcementlearningwildfiremitigation,
+      title={Reinforcement Learning for Wildfire Mitigation in Simulated Disaster Environments}, 
+      author={Alexander Tapley and Marissa Dotter and Michael Doyle and Aidan Fennelly and Dhanuj Gandikota and Savanna Smith and Michael Threet and Tim Welsh},
+      year={2023},
+      eprint={2311.15925},
+      archivePrefix={arXiv},
+      primaryClass={cs.LG},
+      url={https://arxiv.org/abs/2311.15925}, 
+}
+
+@article{KULIGOWSKI2021103129,
+title = {Evacuation decision-making and behavior in wildfires: Past research, current challenges and a future research agenda},
+journal = {Fire Safety Journal},
+volume = {120},
+pages = {103129},
+year = {2021},
+note = {Fire Safety Science: Proceedings of the 13th International Symposium},
+issn = {0379-7112},
+doi = {https://doi.org/10.1016/j.firesaf.2020.103129},
+url = {https://www.sciencedirect.com/science/article/pii/S0379711220302204},
+author = {Erica Kuligowski},
+keywords = {Wildfires, Human behavior, Evacuation, Modeling, Bushfires, WUI fires},
+abstract = {Wildfires are becoming more common around the world, and households are frequently advised to evacuate when these fires threaten nearby communities. Effective evacuation requires an understanding of human behavior in wildfires, which is an area that needs further exploration. The purpose of this article is to present current research performed and data collected on evacuation decision-making and behavior during wildland-urban interface (WUI) fires, identify gaps in the research, and develop a future research plan for further data collection of important WUI fire evacuation topics. Research in this area can support developments of evacuation simulation models, and improvements in education programs, planning, decision-making, and design requirements for community-wide WUI fire evacuation.}
+}
+
+@article{WANG201686,
+title = {An agent-based model of a multimodal near-field tsunami evacuation: Decision-making and life safety},
+journal = {Transportation Research Part C: Emerging Technologies},
+volume = {64},
+pages = {86-100},
+year = {2016},
+issn = {0968-090X},
+doi = {https://doi.org/10.1016/j.trc.2015.11.010},
+url = {https://www.sciencedirect.com/science/article/pii/S0968090X15004106},
+author = {Haizhong Wang and Alireza Mostafizi and Lori A. Cramer and Dan Cox and Hyoungsu Park},
+keywords = {Agent-based modeling, Tsunami evacuation, Multimodal, Decision-making, Life safety},
[…]
