+++ paper/paper.md
@@ -32,11 +32,12 @@ A major effect of climate change today is the increased frequency and intensity 
 
 # Statement of Need
 
-There has been significant traction in the use of computational methods to study wildfires. In particular, reinforcement learning -- a subdomain of artificial intelligence where models learn through interaction with their environment -- has seen growing interest from researchers. Applying reinforcement learning requires modeling the spread of wildfires. Traditionally, modeling was primarily done using physics-based methods [@rothermel1972mathematical; @Andrews_1986]. However, newer methods are more data-driven, enabling the use of a higher diversity of features [@https://doi.org/10.1002/eap.1898; @Diao2020].
+There has been significant traction in the use of computational models to study wildfires. Historically, much work has focused on accurately modeling the spread of wildfires. While a lot of older methods were primarily done using physics-based methods  [@rothermel1972mathematical; @Andrews_1986] – with Rothermel being one of the most popular, as well the one we utilize in our package – newer methods rely on machine learning and other data-driven approaches, incorporating a higher diversity of features  [@https://doi.org/10.1002/eap.1898; @Diao2020;@ross2021being].
 
-Researchers have recently been studying wildfire surveillance and monitoring. While various forms of machine learning, such as computer vision [@ganapathi2018using], have been used to solve this task, the most popular method by far has been reinforcement learning [@Julian2019; @altamimi2022large; @9340340]. Research in surveillance and monitoring has been supported by open-source environments for modeling wildfire spread and surveillance [@cellular_automata; @forest_fire].
+Reinforcement learning (RL), a subdomain of artificial intelligence where models learn through interaction with their environment – has also been increasingly used in the context of wildfires. In combination with other traditional statistical methods and computer vision [@ganapathi2018using; @satelliteimages2017], RL has been applied to both the surveillance and monitoring of wildfires [@Julian2019; @altamimi2022large; @9340340]. An area where there has been little work in regards to RL is wildfire evacuation. Understanding the effective approaches for evacuating populated areas during wildfires is a key safety concern during these events [@KULIGOWSKI2021103129, @McCaffrey_2017]. As a result, work has been done to better model traffic during wildfire evacuation scenarios [@Pel, @doi:10.1061/JTEPBS.0000221], and agent-based evacuation simulations have been used for not only wildfires but also other natural disasters like tsunamis [@BELOGLAZOV2016144, @WANG201686]. RL has been previously identified as a potentially helpful tool during evacuation operations [@rempel_shiell_2022] and has been used to model evacuation during electrical substation fires [@10.1063/5.0209018]. The application of RL techniques to the wildfire evacuation task could thus prove beneficial. 
 
-There has also been emerging interest in optimizing the evacuation process using computational methods [@https://doi.org/10.1111/risa.12944]. However, no reinforcement learning environments exist for the task of evacuation. We hope that open-source tools for evacuation will spur development in this area.
+Given the growing interest in studying wildfires through a computational lens, there have been developments in simulators for wildfires. A lot of open-source software focus on providing a visualization of wildfire spread [@cellular_automata; @forest_fire]. The most relevant piece of work to our paper are SimFire and SimHarness, which provide a system for wildland fire spread and a way for appropriate mitigation strategy responses via RL [@tapley2023reinforcementlearningwildfiremitigation]. Nonetheless, the focus is still on wildfire surveillance and mitigation, not on the task of evacuation.
+
 
 # Methods
 
