+++ paper/paper.bib
@@ -165,6 +165,14 @@ url={https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conf
     eprint = {https://pubs.aip.org/aip/acp/article-pdf/doi/10.1063/5.0209018/19898413/040005\_1\_5.0209018.pdf},
 }
 
+@inproceedings{inproceedings,
+author = {Subramanian, Sriram and Crowley, Mark},
+year = {2017},
+month = {01},
+pages = {},
+title = {Learning Forest Wildfire Dynamics from Satellite Images Using Reinforcement Learning}
+}
+
 @InProceedings{Diao2020,
 author = {Tina Diao and Samriddhi Singla and Ayan Mukhopadhyay and Ahmed Eldawy and Ross Shachter and Mykel J. Kochenderfer},
 booktitle = {AIAA Fall Symposium},
