+++ paper/paper.bib
@@ -9,6 +9,16 @@
     doi = {10.5281/zenodo.8127026},
 }
 
+@article{firetech,
+author = {Xu, Ningzhe and Lovreglio, Ruggiero and Kuligowski, Erica and Cova, Thomas and Nilsson, Daniel and Zhao, Xilei},
+year = {2023},
+month = {01},
+pages = {},
+title = {Predicting and Assessing Wildfire Evacuation Decision-Making Using Machine Learning: Findings from the 2019 Kincade Fire},
+journal = {Fire Technology},
+doi = {10.1007/s10694-023-01363-1}
+}
+
 @misc{tapley2023reinforcementlearningwildfiremitigation,
       title={Reinforcement Learning for Wildfire Mitigation in Simulated Disaster Environments}, 
       author={Alexander Tapley and Marissa Dotter and Michael Doyle and Aidan Fennelly and Dhanuj Gandikota and Savanna Smith and Michael Threet and Tim Welsh},
