### 2024-08-14 sisl/PyroRL#13 @joey-obrien
Hey @SamTov - thanks for the feedback! We reworked the related work in the Statement of Need to include four of the five references you listed above alongside more references to work done on wildfire evacuation and work that used RL for evacuation.
