### 2024-08-14 #6739 @joey-obrien
Hey @mikemahoney218 - we have updated our related works in our Statement of Need as @SamTov requested in sisl/PyroRL/issues/13 and the paper should now be finished. @cpondoc has addressed everything in sisl/PyroRL/issues/14 and will provide more details on the changes in that thread tonight. He will also look into the fix for the buggy window talked about in sisl/PyroRL/issues/12. Thanks!
