+++ paper/paper.bib
@@ -175,11 +175,12 @@ url={https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conf
     eprint = {https://pubs.aip.org/aip/acp/article-pdf/doi/10.1063/5.0209018/19898413/040005\_1\_5.0209018.pdf},
 }
 
-@inproceedings{inproceedings,
+@inproceedings{satelliteimages2017,
 author = {Subramanian, Sriram and Crowley, Mark},
 year = {2017},
 month = {01},
 pages = {},
+booktitle = {The 3rd Multi-disciplinary Conference on Reinforcement Learning and Decision Making},
 title = {Learning Forest Wildfire Dynamics from Satellite Images Using Reinforcement Learning}
 }
 
@@ -190,6 +191,14 @@ title = {Uncertainty aware wildfire management},
 year = {2020},
 url = {https://arxiv.org/pdf/2010.07915.pdf},
 doi = {10.48550/arXiv.2010.07915}
+}
+
+@inproceedings{ross2021being,
+  title={Being the Fire: A CNN-Based Reinforcement Learning Method to Learn How Fires Behave Beyond the Limits of Physics-Based Empirical Models},
+  author={Ross, William L},
+  booktitle={NeurIPS 2021 Workshop on Tackling Climate Change with Machine Learning},
+  url={https://www.climatechange.ai/papers/neurips2021/22},
+  year={2021}
 }
 
 @article{ganapathi2018using,
