### 2024-08-14 #6739 @joey-obrien
I also did want to quickly ask about one of our citations - we use an article that is originally from a NATO conference but was published in a NATO journal later (after it was declassified it seems). Should we cite it as coming from the journal or the conference (we do the conference right now):https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conference-2022/68-using-reinforcement-learning-to-provide-decision-support-in-multi-domain-mass-evacuation-operations Thanks!
