+++ paper/paper.md
@@ -32,7 +32,7 @@ A major effect of climate change today is the increased frequency and intensity 
 
 # Statement of Need
 
-There has been significant traction in the use of computational models to study wildfires. Historically, much work has focused on accurately modeling the spread of wildfires. While a lot of older methods were primarily done using physics-based methods  [@rothermel1972mathematical; @Andrews_1986] – with Rothermel being one of the most popular, as well the one we utilize in our package – newer methods rely on machine learning and other data-driven approaches, incorporating a higher diversity of features  [@https://doi.org/10.1002/eap.1898; @Diao2020;@ross2021being].
+There has been significant traction in the use of computational models to study wildfires. Historically, much work has focused on accurately modeling the spread of wildfires. While a lot of older methods were primarily done using physics-based methods  [@rothermel1972mathematical; @Andrews_1986] – with Rothermel being one of the most popular, as well the one we utilize in our package – newer methods rely on machine learning and other data-driven approaches, incorporating a higher diversity of features  [@https://doi.org/10.1002/eap.1898; @Diao2020; @ross2021being].
 
 Reinforcement learning (RL), a subdomain of artificial intelligence where models learn through interaction with their environment – has also been increasingly used in the context of wildfires. In combination with other traditional statistical methods and computer vision [@ganapathi2018using; @satelliteimages2017], RL has been applied to both the surveillance and monitoring of wildfires [@Julian2019; @altamimi2022large; @9340340]. An area where there has been little work in regards to RL is wildfire evacuation. Understanding the effective approaches for evacuating populated areas during wildfires is a key safety concern during these events [@KULIGOWSKI2021103129; @McCaffrey_2017], and other machine learning techniques have proven beneficial for evacuation planning [@firetech]. As a result, work has been done to better model traffic during wildfire evacuation scenarios [@Pel; @doi:10.1061/JTEPBS.0000221], and agent-based evacuation simulations have been used for not only wildfires but also other natural disasters like tsunamis [@BELOGLAZOV2016144; @WANG201686]. RL has been previously identified as a potentially helpful tool during evacuation operations [@rempel_shiell_2022] and has been used to model evacuation during electrical substation fires [@10.1063/5.0209018]. The application of RL techniques to the wildfire evacuation task could thus prove beneficial. 
 
