### 2024-08-14 #6739 @cpondoc
As mentioned by @joey-obrien -- paper content itself should address reviewer comments. We will also be able to fix DOIs. Investigating the window issue and can come up with patch by tomorrow but shouldn't be blocking from the paper review.
