### 2024-08-14 #6739 @SamTov
@cpondoc The window issue was just a bug I found playing around with the code so I don't even think it should block publication. But knowing about bugs is always helpful.
