### 2024-08-14 #6739 @mikemahoney218
Thanks @cpondoc and @joey-obrien .

@SamTov and @shahchiragh , would you be able to check the authors' responses to your comments and either follow up with additional feedback or, if your initial concerns were satisfied, comment here saying that you don't have any other concerns?

@joey-obrien , with regards to:
> we use an article that is originally from a NATO conference but was published in a NATO journal later

I don't think it matters that much. My rough rule of thumb is "cite the version you actually read", but if you'd rather change the citation to the journal version, that's also perfectly fine.
