### 2024-08-15 431e753
Fix invalid DOIs
