+++ paper/paper.bib
@@ -1,299 +1,302 @@
 @misc{towers_gymnasium_2023,
-    title = {Gymnasium},
-    url = {https://zenodo.org/record/8127025},
-    abstract = {An API standard for single-agent reinforcement learning environments, with popular reference environments and related utilities (formerly Gym)},
-    urldate = {2023-07-08},
-    author = {Towers, Mark and Terry, Jordan K. and Kwiatkowski, Ariel and Balis, John U. and Cola, Gianluca de and Deleu, Tristan and Goulão, Manuel and Kallinteris, Andreas and KG, Arjun and Krimmel, Markus and Perez-Vicente, Rodrigo and Pierré, Andrea and Schulhoff, Sander and Tai, Jun Jet and Shen, Andrew Tan Jin and Younis, Omar G.},
-    month = mar,
-    year = {2023},
-    doi = {10.5281/zenodo.8127026},
+  title    = {Gymnasium},
+  url      = {https://zenodo.org/record/8127025},
+  abstract = {An API standard for single-agent reinforcement learning environments, with popular reference environments and related utilities (formerly Gym)},
+  urldate  = {2023-07-08},
+  author   = {Towers, Mark and Terry, Jordan K. and Kwiatkowski, Ariel and Balis, John U. and Cola, Gianluca de and Deleu, Tristan and Goulão, Manuel and Kallinteris, Andreas and KG, Arjun and Krimmel, Markus and Perez-Vicente, Rodrigo and Pierré, Andrea and Schulhoff, Sander and Tai, Jun Jet and Shen, Andrew Tan Jin and Younis, Omar G.},
+  month    = mar,
+  year     = {2023},
+  doi      = {10.5281/zenodo.8127026}
 }
 
 @article{firetech,
-author = {Xu, Ningzhe and Lovreglio, Ruggiero and Kuligowski, Erica and Cova, Thomas and Nilsson, Daniel and Zhao, Xilei},
-year = {2023},
-month = {01},
-pages = {},
-title = {Predicting and Assessing Wildfire Evacuation Decision-Making Using Machine Learning: Findings from the 2019 Kincade Fire},
-journal = {Fire Technology},
-doi = {10.1007/s10694-023-01363-1}
+  author  = {Xu, Ningzhe and Lovreglio, Ruggiero and Kuligowski, Erica and Cova, Thomas and Nilsson, Daniel and Zhao, Xilei},
+  year    = {2023},
+  month   = {01},
+  pages   = {},
+  title   = {Predicting and Assessing Wildfire Evacuation Decision-Making Using Machine Learning: Findings from the 2019 Kincade Fire},
+  journal = {Fire Technology},
+  doi     = {10.1007/s10694-023-01363-1}
 }
 
 @misc{tapley2023reinforcementlearningwildfiremitigation,
-      title={Reinforcement Learning for Wildfire Mitigation in Simulated Disaster Environments}, 
-      author={Alexander Tapley and Marissa Dotter and Michael Doyle and Aidan Fennelly and Dhanuj Gandikota and Savanna Smith and Michael Threet and Tim Welsh},
-      year={2023},
-      eprint={2311.15925},
-      archivePrefix={arXiv},
-      primaryClass={cs.LG},
-      url={https://arxiv.org/abs/2311.15925}, 
+  title         = {Reinforcement Learning for Wildfire Mitigation in Simulated Disaster Environments},
+  author        = {Alexander Tapley and Marissa Dotter and Michael Doyle and Aidan Fennelly and Dhanuj Gandikota and Savanna Smith and Michael Threet and Tim Welsh},
+  year          = {2023},
+  eprint        = {2311.15925},
+  archiveprefix = {arXiv},
+  primaryclass  = {cs.LG},
[…]
