+++ paper/paper.bib
@@ -184,6 +184,7 @@
   month     = {01},
   pages     = {},
   booktitle = {The 3rd Multi-disciplinary Conference on Reinforcement Learning and Decision Making},
+  doi       = {10.3389/fict.2018.00006},
   title     = {Learning Forest Wildfire Dynamics from Satellite Images Using Reinforcement Learning}
 }
 
