### 2024-08-15 #6739 @cpondoc
Hey @mikemahoney218  -- thanks for circling back! Also wanted to address the last couple of DOIs. After doing a diff, the only ones left missing were:
```
- No DOI given, and none found for title: Reinforcement Learning for Wildfire Mitigation in ...
- No DOI given, and none found for title: Using Reinforcement Learning to Provide Decision S...
- No DOI given, and none found for title: Learning Forest Wildfire Dynamics from Satellite I...
- No DOI given, and none found for title: Being the Fire: A CNN-Based Reinforcement Learning...
```

The first and the last papers are both workshop papers with no DOIs, the second is the NATO paper, and I should've just found the DOI for the third paper. Think that just about covers things from our end.
