### 2024-09-04 #6739 @mikemahoney218
@cpondoc , the paper is looking good! I've got a few changes to make in the references section:

- Can you remove the inner braces here, so our CSL gets applied? https://github.com/sisl/PyroRL/blob/master/paper/paper.bib#L232
- Can you add authors to Gym Cellular Automata ( https://github.com/elbecerrasoto/gym-cellular-automata/blob/main/AUTHORS.md ) and Gym Forest Fire
- I don't know _why_ it's happening, but the word "I" is lowercase in McCaffrey 2017's citation in the PDF. Can you maybe put `{}` around those in the title (unless you see the real reason it's happening?) https://github.com/sisl/PyroRL/blob/master/paper/paper.bib#L105

Can you also go through the "Author Tasks" checklist here: https://github.com/openjournals/joss-reviews/issues/6739#issuecomment-2327595662 and post the links to your archive and release when finished?
