### 2024-09-05 #6739 @cpondoc
Hey @mikemahoney218 -- this is awesome! @joey-obrien will be able to take care of the paper edits. Here is the release [link](https://github.com/sisl/PyroRL/releases/tag/v1.0.0). Out of curiosity, do you have a list of steps on how to archive the release on Zenodo/figshare/etc and generate the DOI?
