### 2024-09-11 #6739 @cpondoc
Hey @mikemahoney218 -- sounds good! @joey-obrien will get to the paper edits later today EST. For the GitHub release: did you check out this [link](https://github.com/sisl/PyroRL/releases/tag/v1.0.0)? Thanks!
