### 2024-09-12 1bfbac1
paper bibliography updates
