+++ paper/paper.bib
@@ -102,7 +102,7 @@
 }
 
 @article{McCaffrey_2017,
-  title     = {Should I Stay or Should I Go Now? Or Should I Wait and See? Influences on Wildfire Evacuation Decisions},
+  title     = {Should {I} Stay or Should {I} Go Now? Or Should {I} Wait and See? Influences on Wildfire Evacuation Decisions},
   volume    = {38},
   issn      = {1539-6924},
   url       = {http://dx.doi.org/10.1111/risa.12944},
@@ -152,14 +152,12 @@
   year     = {2019}
 }
 
-@inproceedings{rempel_shiell_2022,
+@inproceedings{rempel_shiell_2023,
   title     = {Using Reinforcement Learning to Provide Decision Support in Multi-Domain Mass Evacuation Operations},
   url       = {https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conference-2022/68-using-reinforcement-learning-to-provide-decision-support-in-multi-domain-mass-evacuation-operations},
-  booktitle = {NATO Operations Research and Analysis Conference},
+  journal   = {NATO STO Review},
   author    = {Rempel, Mark and Shiell, Nicholi},
-  year      = {2022},
-  month     = {October},
-  address   = {Copenhagen, Denmark}
+  year      = {2023}
 }
 
  @article{10.1063/5.0209018,
@@ -229,7 +227,7 @@
 }
 
 @article{altamimi2022large,
-  title     = {{Large-Scale Wildfire Mitigation Through Deep Reinforcement Learning}},
+  title     = {Large-Scale Wildfire Mitigation Through Deep Reinforcement Learning},
   author    = {Altamimi, Abdulelah and Lagoa, Constantino and Borges, Jos{\'e} G and McDill, Marc E and Andriotis, CP and Papakonstantinou, KG},
   journal   = {Frontiers in Forests and Global Change},
   volume    = {5},
@@ -266,12 +264,14 @@
 }
 
 @software{cellular_automata,
+  author = {Emanuel Becerra Soto},
   title = {{Gym Cellular Automata}},
   year  = {2021},
   url   = {https://github.com/elbecerrasoto/gym-cellular-automata}
 }
 
 @software{forest_fire,
+  author = {Sahand Rezaei-Shoshtari},
   title = {{Gym Forest Fire}},
   year  = {2020},
   url   = {https://github.com/sahandrez/gym_forestfire}
