+++ paper/paper.bib
@@ -300,7 +300,7 @@
   isbn      = {0262047012}
 }
 
-@article{stableBaselines,
+@article{stable_baselines,
   author  = {Antonin Raffin and Ashley Hill and Adam Gleave and Anssi Kanervisto and Maximilian Ernestus and Noah Dormann},
   title   = {Stable-Baselines3: Reliable Reinforcement Learning Implementations},
   journal = {Journal of Machine Learning Research},
