### 2024-09-12 #6739 @joey-obrien
Hey @mikemahoney218, I made the changes you referred to above in the references. I also updated the reference to the NATO article to be from the journal I found it in and not the conference. The URL for this NATO article also currently extends beyond the edge of the page, and I was wondering if there is something we should do to fix this. Thanks!
