### 2024-09-16 #6739 @crvernon
@cpondoc - very nice job with this one.  I only have one comment that needs to be addressed before I accept this for publication.

- [ ] LINE 189:  "kincade fire" should have a capital "K" and "F" in this context.  You can maintain formatting in your bib file using curly brackets around the characters that you wish to do so for.

Let me know when this has been addressed and I'll publish.
