+++ paper/paper.bib
@@ -14,7 +14,7 @@
   year    = {2023},
   month   = {01},
   pages   = {},
-  title   = {Predicting and Assessing Wildfire Evacuation Decision-Making Using Machine Learning: Findings from the 2019 Kincade Fire},
+  title   = {Predicting and Assessing Wildfire Evacuation Decision-Making Using Machine Learning: Findings from the 2019 {K}incade {F}ire},
   journal = {Fire Technology},
   doi     = {10.1007/s10694-023-01363-1}
 }
@@ -153,11 +153,11 @@
 }
 
 @inproceedings{rempel_shiell_2023,
-  title     = {Using Reinforcement Learning to Provide Decision Support in Multi-Domain Mass Evacuation Operations},
-  url       = {https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conference-2022/68-using-reinforcement-learning-to-provide-decision-support-in-multi-domain-mass-evacuation-operations},
-  journal   = {NATO STO Review},
-  author    = {Rempel, Mark and Shiell, Nicholi},
-  year      = {2023}
+  title   = {Using Reinforcement Learning to Provide Decision Support in Multi-Domain Mass Evacuation Operations},
+  url     = {https://review.sto.nato.int/index.php/journal-issues/2023-fall/sas-ora-conference-2022/68-using-reinforcement-learning-to-provide-decision-support-in-multi-domain-mass-evacuation-operations},
+  journal = {NATO STO Review},
+  author  = {Rempel, Mark and Shiell, Nicholi},
+  year    = {2023}
 }
 
  @article{10.1063/5.0209018,
@@ -265,16 +265,16 @@
 
 @software{cellular_automata,
   author = {Emanuel Becerra Soto},
-  title = {{Gym Cellular Automata}},
-  year  = {2021},
-  url   = {https://github.com/elbecerrasoto/gym-cellular-automata}
+  title  = {{Gym Cellular Automata}},
+  year   = {2021},
+  url    = {https://github.com/elbecerrasoto/gym-cellular-automata}
 }
 
 @software{forest_fire,
   author = {Sahand Rezaei-Shoshtari},
-  title = {{Gym Forest Fire}},
-  year  = {2020},
-  url   = {https://github.com/sahandrez/gym_forestfire}
+  title  = {{Gym Forest Fire}},
+  year   = {2020},
+  url    = {https://github.com/sahandrez/gym_forestfire}
 }
 
 @article{Julian2019,
