### 2025-05-30 #6739 @arfon
Looks like https://doi.org/10.21105/joss.06739 is now resolving.
